Add explicit Manifest2822Maintainer struct
This allows us to more easily/reliably generate lists of users to ping on PRs (with the goal to ping them automatically, if one of them is not the PR author).

This is technically a breaking change, but I consider myself probably the most prolific user of bashbrew format strings (which is where this breakage would exhibit itself -- `bashbrew cat` should be unaffected), and I have exactly one script broken by it.  That script was doing some processing to generate this exact data, so IMO this is a win. 😁

    bashbrew cat --format='- `{{ $.RepoName }}`:{{ range .Manifest.Global.Maintainers }} @{{ .Handle }}{{ end }}' ...